# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WIZIA API Documentation",
        "description": "Documentation de l'API WIZIA",
        "contact": {
            "email": "contact@wizia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Récupérer un utilisateur par ID",
                "operationId": "a5b1c36dde931f51716ec1e224d446b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "797ddc9b6cbdfa76d0078b06e2d0b7c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johnexample.com"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "0601020304"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "activity": {
                                        "type": "string",
                                        "example": "coiffeur"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "description"
                                    },
                                    "companyName": {
                                        "type": "string",
                                        "example": "WIZIA"
                                    },
                                    "tone": {
                                        "type": "string",
                                        "example": "Familier"
                                    },
                                    "goal": {
                                        "type": "string",
                                        "example": "Vendre mes produits"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "4a096d5472738a3fd2489d43dadb76ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/users/sertchUser": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Rechercher un utilisateur via email et mot de passe",
                "operationId": "a84269cc2b71053f5eed937e5f8f8d64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "userexample.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé et authentifié"
                    },
                    "401": {
                        "description": "Email ou mot de passe invalide"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Ajouter un utilisateur",
                "operationId": "f1937fb83c59be9492411ceb08dc859b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "activity",
                                    "password",
                                    "password_confirmation",
                                    "color",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johnexample.com"
                                    },
                                    "activity": {
                                        "type": "string",
                                        "example": "Coiffure"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0601020304"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "example": "data:image/png;base64,..."
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#FF5733"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description du profil"
                                    },
                                    "tone": {
                                        "type": "string",
                                        "example": "Familier"
                                    },
                                    "goal": {
                                        "type": "string",
                                        "example": "Vendre mes produits"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/auth/AuthenticatedUser": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Récupérer l'utilisateur actuellement authentifié",
                "operationId": "88880e2b69fabf5f37a21175fef11e36",
                "responses": {
                    "200": {
                        "description": "Utilisateur authentifié retourné"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Authentifier un utilisateur",
                "operationId": "0a12ad9483755caabdc23f730286aa9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johnexample.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie, retourne token et user"
                    },
                    "401": {
                        "description": "Email ou mot de passe invalide"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/users/uploadlogo": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Uploader l'image du logo de l'utilisateur",
                "operationId": "a6313fb76434cb176f5be4e23a10501e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Image à uploader (jpeg,png,jpg,gif)",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image uploadée avec succès"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/users/abonnementUser": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Récupérer les informations d'abonnement et limites d'un utilisateur",
                "operationId": "e84934423b773036029bd0f5a187c4eb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Infos d'abonnement récupérées"
                    },
                    "404": {
                        "description": "Utilisateur ou abonnement non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        }
    },
...
}